# Endpoint docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ShopifyApp",
    "version": "1.0"
  },
  "paths": {
    "/api/Customers/redeem": {
      "post": {
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRedeemModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRedeemModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRedeemModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GiftCard"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiftCard"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiftCard"
                }
              }
            }
          }
        }
      }
    },
    "/api/Customers/points/{customerId}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AvailablePointsModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailablePointsModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailablePointsModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailablePointsModel": {
        "type": "object",
        "properties": {
          "CustomerId": {
            "type": "string",
            "nullable": true
          },
          "AvailablePoints": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomerRedeemModel": {
        "required": [
          "CustomerId",
          "Points"
        ],
        "type": "object",
        "properties": {
          "CustomerId": {
            "type": "integer",
            "format": "int64"
          },
          "Points": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GiftCard": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "AdminGraphQLAPIId": {
            "type": "string",
            "nullable": true
          },
          "Balance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Currency": {
            "type": "string",
            "nullable": true
          },
          "InitialValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "DisabledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ExpiresOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "LineItemId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "ApiClientId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "UserId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "CustomerId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Note": {
            "type": "string",
            "nullable": true
          },
          "TemplateSuffix": {
            "type": "string",
            "nullable": true
          },
          "LastCharacters": {
            "type": "string",
            "nullable": true
          },
          "OrderId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}